statusbar: show "developer mode" info in tooltip
