+++ package.nls.json
@@ -230,7 +230,6 @@
     "AWS.credentials.statusbar.no.credentials": "(not connected)",
     "AWS.credentials.statusbar.connected": "(connected)",
     "AWS.credentials.statusbar.text": "{0}: {1}",
-    "AWS.credentials.statusbar.tooltip": "The current credentials used by the {0} Toolkit.\n\nClick this status bar item to use different credentials.",
     "AWS.error.during.sam.local": "Failed to run SAM application locally: {0}",
     "AWS.error.endpoint.load.failure": "The {0} Toolkit was unable to load endpoints data.",
     "AWS.error.impactedFunctionalityReset.vscode": "Toolkit functionality may be impacted until VS Code is restarted.",
