+++ src/test/utilities/fakeAwsContext.ts
@@ -70,7 +70,7 @@ export class FakeAwsContext implements AwsContext {
         this.awsContextCredentials = params?.contextCredentials
     }
 
-    public async setDeveloperMode(enable: boolean): Promise<void> {}
+    public async setDeveloperMode(enable: boolean, settingName: string | undefined): Promise<void> {}
 
     public async setCredentials(credentials?: AwsContextCredentials): Promise<void> {
         this.awsContextCredentials = credentials
